+++ src/index.ts
@@ -1,2 +1,2 @@
 export { RapidServer } from "./server/RapidServer"
-export { getRapidObjectClass } from "./builder/RapidClassBuilder"
+export { getRapidObjectClass } from "./builder/rapid-class.builder"