+++ src/interfaces/IEndpoint.ts
@@ -1,4 +1,0 @@
-export interface Endpoint {
-    name: string,
-    method: "GET" | "POST" | "PUT" | "PATCH" | "DELETE"
-}