+++ src/interfaces/IRapidEndpoint.ts
@@ -1,6 +1,0 @@
-export interface RapidEndpoint {
-    name: string,
-    methods: "GET" | "POST" | "PUT" | "PATCH" | "DELETE" | "GET_BY_ID"[],
-    object: string,
-    hasId: boolean
-}