+++ src/types/allowed-http-methods.d.ts
@@ -1,0 +1,1 @@
+export type AllowedHttpMethods = "GET" | "POST" | "PUT" | "DELETE" | "PATCH";