+++ src/types/express-endpoint.d.ts
@@ -1,0 +1,6 @@
+import {AllowedHttpMethods} from "./allowed-http-methods";
+
+export type ExpressEndpoint = {
+    name: string,
+    method: AllowedHttpMethods
+}