+++ src/types/html-endpoint.d.ts
@@ -1,4 +1,4 @@
-export interface EndpointHTML {
+export type HtmlEndpoint = {
     methods: string[],
     name: string,
 }