+++ src/types/rapid-config.d.ts
@@ -1,0 +1,8 @@
+import {RapidEndpoint} from "../interfaces/IRapidEndpoint";
+
+export type RapidConfig = {
+    name: string,
+    endpoints: RapidEndpoint[],
+    port?: number,
+    prefix?: string
+}