+++ src/types/rapid-endpoints.d.ts
@@ -1,0 +1,8 @@
+import {RapidHttpMethods} from "./rapid-http-methods";
+
+export type RapidEndpoint = {
+    name: string,
+    methods: RapidHttpMethods[],
+    object: string,
+    hasId: boolean
+}