+++ src/types/rapid-http-methods.d.ts
@@ -1,0 +1,3 @@
+import {AllowedHttpMethods} from "./http-methods";
+
+export type RapidHttpMethods = AllowedHttpMethods | "GET_BY_ID";