+++ changelog.md
@@ -1,8 +1,11 @@
 ### 1.0.0
-- readme yaay
+- get rid of config.json and add config object to RapidServer constructor
+- massive refactoring
+- fixed wrong error message when wrogn data type is provided
+- added error handling for invalid property names
+- updated readme
 
 ### 0.0.9
-- added build scripts
 - code refactoring
 - readme updates